requirements.txt
# This file is generated automatically by `make requirements.txt`.
# Do not edit manually.
matplotlib==3.8.0
numpy==1.26.2
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Requests==2.31.0
scipy==1.11.4
trimesh==4.0.5
transforms3d==0.4.1
tqdm
pyyaml
paho-mqtt
utm
scipy
# lap
filterpy
keyboard
